+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Development
 ### 2018-10-23
 * Objects: Don't stamp if the canvas is too small (and would throw an error)
+* Threads: New "reportAtomicGroup" HOF primitive using the JIT compiler
 
 ### 2018-10-22
 * "Bigger Data" library: Added "analyze" function for fast frequency distributions